Fixed circular dependency issue with SecurityTokenService
