+++ src/main/java/no/nav/bidrag/beregn/forskudd/rest/security/OAuthClientConfiguration.java
@@ -1,10 +1,7 @@
-package no.nav.bidrag.beregn.forskudd.rest;
+package no.nav.bidrag.beregn.forskudd.rest.security;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.oauth2.client.AuthorizedClientServiceOAuth2AuthorizedClientManager;
 import org.springframework.security.oauth2.client.OAuth2AuthorizedClientManager;
 import org.springframework.security.oauth2.client.OAuth2AuthorizedClientProviderBuilder;
@@ -12,22 +9,7 @@ import org.springframework.security.oauth2.client.OAuth2AuthorizedClientService;
 import org.springframework.security.oauth2.client.registration.ClientRegistrationRepository;
 
 @Configuration
-class SecurityConfiguration extends WebSecurityConfigurerAdapter {
-
-  @Override
-  public void configure(HttpSecurity http) throws Exception {
-    http.sessionManagement()
-        .sessionCreationPolicy(SessionCreationPolicy.STATELESS)
-        .and()
-        .csrf()
-        .disable()
-        .authorizeRequests()
-        .antMatchers("/**")
-        .permitAll()
-        .anyRequest()
-        .fullyAuthenticated();
-  }
-
+public class OAuthClientConfiguration {
   @Bean
   public OAuth2AuthorizedClientManager authorizedClientManager(
       ClientRegistrationRepository clientRegistrationRepository,
