+++ src/main/java/no/nav/bidrag/beregn/forskudd/rest/security/SecurityConfiguration.java
@@ -1,0 +1,24 @@
+package no.nav.bidrag.beregn.forskudd.rest.security;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.config.http.SessionCreationPolicy;
+
+@Configuration
+class SecurityConfiguration extends WebSecurityConfigurerAdapter {
+
+  @Override
+  public void configure(HttpSecurity http) throws Exception {
+    http.sessionManagement()
+        .sessionCreationPolicy(SessionCreationPolicy.STATELESS)
+        .and()
+        .csrf()
+        .disable()
+        .authorizeRequests()
+        .antMatchers("/**")
+        .permitAll()
+        .anyRequest()
+        .fullyAuthenticated();
+  }
+}
